Introduce DataObject Magic Method extension
Refactor existing code to make sure we can add multiple
magic method reflection extensions for various use cases.
