+++ extension.neon
@@ -4,6 +4,10 @@ services:
 		tags:
 			- phpstan.broker.dynamicMethodReturnTypeExtension
 	-
-		class: bitExpert\PHPStan\Magento\Reflection\SessionManagerMagicMethodReflectionExtension
+		class: bitExpert\PHPStan\Magento\Reflection\Framework\Session\SessionManagerMagicMethodReflectionExtension
+		tags:
+			- phpstan.broker.methodsClassReflectionExtension
+	-
+		class: bitExpert\PHPStan\Magento\Reflection\Framework\DataObjectMagicMethodReflectionExtension
 		tags:
 			- phpstan.broker.methodsClassReflectionExtension
